requirements.txt
flask-mongoengine==0.7.0
eventlet
bcrypt<2.0.
Flask-RESTful
flask-security
flask-socketio
pytest
flask-cors